examples/gcp_health_agent/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.75.0
python-dotenv
google-adk>=0.1.0
google-auth-oauthlib>=1.2.2
google-api-python-client
google-auth
google-cloud-logging
google-cloud-monitoring
google-cloud-run
google-cloud-resource-manager>=1.0.0
google-cloud-securitycenter
google-cloud-recommender
google-cloud-servicehealth>=0.1.0
google-cloud-network-management>=1.0.0